Flatten data structures and add url support
This commit refactors the data structures for managing the Mandelbrot
set calculations and introduces several new features.

The `Surface` class is removed, and the state of each view is again
managed by a set of parallel arrays (`zz`, `cc`, `nn`, etc.). This
change is aimed at improving performance by using more direct data
structures.

This version now supports URL parameters to control the initial view.
The `grid`, `c` (center), and `s` (size) parameters can be used to
link to a specific view of the Mandelbrot set.